✨ feat (CHASE Frontend): Chairs can now generate their own messages for WD branch: 98-enable-chairs-to-create-their-own-messages-for-wd
